Move hostname to SiteHelper module
