+++ app/helpers/application_helper.rb
@@ -51,15 +51,4 @@ module ApplicationHelper
   def sanitize_highlight(text)
     sanitize(text, tags: ["em"])
   end
-
-  # Returns the appropriate [sub] hostname for the current environment
-  sig { params(sub_domain: T.nilable(String)).returns(String) }
-  def hostname(sub_domain = nil)
-    domain = Morph::Application.default_url_options[:host] || "morph.io"
-    if domain.include?("localhost") || sub_domain.blank?
-      domain
-    else
-      "#{sub_domain}.#{domain}"
-    end
-  end
 end
